Break apart large files in probe-rs/debugger/*
