+++ debugger/src/debug_adapter/dap_types.rs
@@ -7,7 +7,7 @@ use schemafy::schemafy;
 use serde::{Deserialize, Serialize};
 use std::convert::TryFrom;
 
-schemafy!(root: debugserver_types "src/debugProtocol.json");
+schemafy!(root: debugserver_types "src/debug_adapter/debugProtocol.json");
 
 /// Custom 'quit' request, so that VSCode can tell the `probe-rs-debugger` to terminate its own process.
 #[derive(Clone, PartialEq, Debug, Deserialize, Serialize)]
