+++ debugger/src/debugger/debug_entry.rs
@@ -1,78 +1,27 @@
-use crate::dap_types::*;
-use crate::debug_adapter::*;
-use crate::protocol::{DapAdapter, ProtocolAdapter};
-
-use crate::DebuggerError;
+use super::session_data;
+use crate::{
+    debug_adapter::{
+        dap_adapter::*,
+        dap_types::*,
+        protocol::{DapAdapter, ProtocolAdapter},
+    },
+    debugger::configuration::{self, ConsoleLog},
+    DebuggerError,
+};
 use anyhow::{anyhow, Context, Result};
-use capstone::{
-    arch::arm::ArchMode as armArchMode, arch::riscv::ArchMode as riscvArchMode, prelude::*,
-    Capstone, Endian,
-};
-use probe_rs::config::TargetSelector;
-use probe_rs::debug::DebugInfo;
-
-use probe_rs::flashing::download_file_with_options;
-use probe_rs::flashing::DownloadOptions;
-use probe_rs::flashing::FlashProgress;
-use probe_rs::flashing::Format;
-use probe_rs::ProbeCreationError;
 use probe_rs::{
-    Core, CoreStatus, DebugProbeError, DebugProbeSelector, Permissions, Probe, Session,
-    WireProtocol,
+    flashing::{download_file_with_options, DownloadOptions, FlashProgress, Format},
+    CoreStatus, Probe,
 };
-use probe_rs_cli_util::rtt;
 use serde::Deserialize;
-use std::cell::RefCell;
-use std::net::Ipv4Addr;
-use std::net::TcpListener;
-use std::ops::Mul;
-use std::rc::Rc;
 use std::{
-    env::{current_dir, set_current_dir},
-    path::PathBuf,
-    str::FromStr,
+    cell::RefCell,
+    net::{Ipv4Addr, TcpListener},
+    ops::Mul,
+    rc::Rc,
     thread,
     time::Duration,
 };
-
-fn default_console_log() -> Option<ConsoleLog> {
-    Some(ConsoleLog::Error)
-}
-
-fn parse_probe_selector(src: &str) -> Result<DebugProbeSelector, String> {
-    match DebugProbeSelector::from_str(src) {
-        Ok(probe_selector) => Ok(probe_selector),
-        Err(error) => Err(error.to_string()),
-    }
-}
-
-/// The level of information to be logged to the debugger console. The DAP Client will set appropriate RUST_LOG env for 'launch' configurations,  and will pass the rust log output to the client debug console.
-#[derive(Copy, Clone, PartialEq, Debug, serde::Serialize, serde::Deserialize)]
-pub enum ConsoleLog {
-    Error,
-    Warn,
-    Info,
-    Debug,
-    Trace,
-}
-
-impl std::str::FromStr for ConsoleLog {
-    type Err = String;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match &s.to_ascii_lowercase()[..] {
-            "error" => Ok(ConsoleLog::Error),
-            "warn" => Ok(ConsoleLog::Error),
-            "info" => Ok(ConsoleLog::Info),
-            "debug" => Ok(ConsoleLog::Debug),
-            "trace" => Ok(ConsoleLog::Trace),
-            _ => Err(format!(
-                "'{}' is not a valid console log level. Choose from [error, warn, info, debug, or trace].",
-                s
-            )),
-        }
-    }
-}
 
 #[derive(clap::Parser, Copy, Clone, Debug, Deserialize)]
 pub(crate) enum TargetSessionType {
@@ -95,550 +44,6 @@ impl std::str::FromStr for TargetSessionType {
     }
 }
 
-/// Shared options for all commands which use a specific probe
-#[derive(clap::Parser, Clone, Deserialize, Debug, Default)]
-#[serde(rename_all = "camelCase")]
-pub struct DebuggerOptions {
-    /// Path to the requested working directory for the debugger
-    #[clap(long, parse(from_os_str))]
-    pub(crate) cwd: Option<PathBuf>,
-
-    /// Binary to debug as a path. Relative to `cwd`, or fully qualified.
-    #[clap(long, parse(from_os_str))]
-    pub(crate) program_binary: Option<PathBuf>,
-
-    /// CMSIS-SVD file for the target. Relative to `cwd`, or fully qualified.
-    #[clap(long, parse(from_os_str))]
-    pub(crate) svd_file: Option<PathBuf>,
-
-    /// The number associated with the debug probe to use. Use 'list' command to see available probes
-    #[clap(
-        long = "probe",
-        parse(try_from_str = parse_probe_selector),
-        help = "Use this flag to select a specific probe in the list.\n\
-                    Use '--probe VID:PID' or '--probe VID:PID:Serial' if you have more than one probe with the same VID:PID."
-    )]
-    #[serde(alias = "probe")]
-    pub(crate) probe_selector: Option<DebugProbeSelector>,
-
-    /// The MCU Core to debug. Default is 0
-    #[clap(long = "core-index", default_value_t)]
-    #[serde(default)]
-    pub(crate) core_index: usize,
-
-    /// The target to be selected.
-    #[clap(short, long)]
-    pub(crate) chip: Option<String>,
-
-    /// Protocol to use for target connection
-    #[clap(short, long)]
-    #[serde(rename = "wire_protocol")]
-    pub(crate) protocol: Option<WireProtocol>,
-
-    /// Protocol speed in kHz
-    #[clap(short, long)]
-    pub(crate) speed: Option<u32>,
-
-    /// Assert target's reset during connect
-    #[clap(long)]
-    #[serde(default)]
-    pub(crate) connect_under_reset: bool,
-
-    /// Allow the chip to be fully erased
-    #[structopt(long)]
-    #[serde(default)]
-    pub(crate) allow_erase_all: bool,
-
-    /// IP port number to listen for incoming DAP connections, e.g. "50000"
-    #[clap(long)]
-    pub(crate) port: Option<u16>,
-
-    /// Flash the target before debugging
-    #[clap(long)]
-    #[serde(default)]
-    pub(crate) flashing_enabled: bool,
-
-    /// Reset the target after flashing
-    #[clap(long, required_if_eq("flashing-enabled", "true"))]
-    #[serde(default)]
-    pub(crate) reset_after_flashing: bool,
-
-    /// Halt the target after reset
-    #[clap(long)]
-    #[serde(default)]
-    pub(crate) halt_after_reset: bool,
-
-    /// Do a full chip erase, versus page-by-page erase
-    #[clap(long, required_if_eq("flashing-enabled", "true"))]
-    #[serde(default)]
-    pub(crate) full_chip_erase: bool,
-
-    /// Restore erased bytes that will not be rewritten from ELF
-    #[clap(long, required_if_eq("flashing-enabled", "true"))]
-    #[serde(default)]
-    pub(crate) restore_unwritten_bytes: bool,
-
-    /// Level of information to be logged to the debugger console (Error, Info or Debug )
-    #[clap(long)]
-    #[serde(default = "default_console_log")]
-    pub(crate) console_log_level: Option<ConsoleLog>,
-
-    #[clap(flatten)]
-    #[serde(flatten)]
-    pub(crate) rtt: rtt::RttConfig,
-}
-
-impl DebuggerOptions {
-    /// Validate the new cwd, or else set it from the environment.
-    pub(crate) fn validate_and_update_cwd(&mut self, new_cwd: Option<PathBuf>) {
-        self.cwd = match new_cwd {
-            Some(temp_path) => {
-                if temp_path.is_dir() {
-                    Some(temp_path)
-                } else if let Ok(current_dir) = current_dir() {
-                    Some(current_dir)
-                } else {
-                    log::error!("Cannot use current working directory. Please check existence and permissions.");
-                    None
-                }
-            }
-            None => {
-                if let Ok(current_dir) = current_dir() {
-                    Some(current_dir)
-                } else {
-                    log::error!("Cannot use current working directory. Please check existence and permissions.");
-                    None
-                }
-            }
-        };
-    }
-
-    /// If the path to the program to be debugged is relative, we join if with the cwd.
-    pub(crate) fn qualify_and_update_os_file_path(
-        &mut self,
-        os_file_to_validate: Option<PathBuf>,
-    ) -> Result<PathBuf, DebuggerError> {
-        match os_file_to_validate {
-            Some(temp_path) => {
-                let mut new_path = PathBuf::new();
-                if temp_path.is_relative() {
-                    if let Some(cwd_path) = self.cwd.clone() {
-                        new_path.push(cwd_path);
-                    } else {
-                        return Err(DebuggerError::Other(anyhow!(
-                            "Invalid value {:?} for `cwd`",
-                            self.cwd
-                        )));
-                    }
-                }
-                new_path.push(temp_path);
-                Ok(new_path)
-            }
-            None => Err(DebuggerError::Other(anyhow!("Missing value for file."))),
-        }
-    }
-}
-
-/// The supported breakpoint types
-#[derive(Debug, PartialEq)]
-pub enum BreakpointType {
-    InstructionBreakpoint,
-    SourceBreakpoint,
-}
-
-/// Provide the storage and methods to handle various [`BreakPointType`]
-#[derive(Debug)]
-pub struct ActiveBreakpoint {
-    breakpoint_type: BreakpointType,
-    breakpoint_address: u32,
-}
-
-/// DebugSession is designed to be similar to [probe_rs::Session], in as much that it provides handles to the [CoreData] instances for each of the available [probe_rs::Core] involved in the debug session.
-/// To get access to the [CoreData] for a specific [Core], the
-/// TODO: Adjust [DebuggerOptions] to allow multiple cores (and if appropriate, their binaries) to be specified.
-pub struct DebugSession {
-    pub(crate) session: Session,
-    /// Provides ability to disassemble binary code.
-    pub(crate) capstone: Capstone,
-    /// [DebugSession] will manage one [DebugInfo] per [DebuggerOptions::program_binary]
-    pub(crate) debug_infos: Vec<DebugInfo>,
-    /// [DebugSession] will manage a `Vec<StackFrame>` per [Core]. Each core's collection of StackFrames will be recreated whenever a stacktrace is performed, using the results of [DebugInfo::unwind]
-    pub(crate) stack_frames: Vec<Vec<probe_rs::debug::StackFrame>>,
-    /// [DebugSession] will manage a `Vec<ActiveBreakpoint>` per [Core]. Each core's collection of ActiveBreakpoint's will be managed on demand.
-    pub(crate) breakpoints: Vec<Vec<ActiveBreakpoint>>,
-    /// The control structures for handling RTT in this Core of the DebugSession.
-    pub(crate) rtt_connection: Option<RttConnection>,
-}
-
-impl DebugSession {
-    pub(crate) fn new(debugger_options: &DebuggerOptions) -> Result<Self, DebuggerError> {
-        let mut target_probe = match debugger_options.probe_selector.clone() {
-            Some(selector) => Probe::open(selector.clone()).map_err(|e| match e {
-                DebugProbeError::ProbeCouldNotBeCreated(ProbeCreationError::NotFound) => {
-                    DebuggerError::Other(anyhow!(
-                        "Could not find the probe_selector specified as {:04x}:{:04x}:{:?}",
-                        selector.vendor_id,
-                        selector.product_id,
-                        selector.serial_number
-                    ))
-                }
-                other_error => DebuggerError::DebugProbe(other_error),
-            }),
-            None => {
-                // Only automatically select a probe if there is only a single probe detected.
-                let list = Probe::list_all();
-                if list.len() > 1 {
-                    return Err(DebuggerError::Other(anyhow!(
-                        "Found multiple ({}) probes",
-                        list.len()
-                    )));
-                }
-
-                if let Some(info) = list.first() {
-                    Probe::open(info).map_err(DebuggerError::DebugProbe)
-                } else {
-                    return Err(DebuggerError::Other(anyhow!(
-                        "No probes found. Please check your USB connections."
-                    )));
-                }
-            }
-        }?;
-
-        let target_selector = match &debugger_options.chip {
-            Some(identifier) => identifier.into(),
-            None => TargetSelector::Auto,
-        };
-
-        // Set the protocol, if the user explicitly selected a protocol. Otherwise, use the default protocol of the probe.
-        if let Some(protocol) = debugger_options.protocol {
-            target_probe.select_protocol(protocol)?;
-        }
-
-        // Set the speed.
-        if let Some(speed) = debugger_options.speed {
-            let actual_speed = target_probe.set_speed(speed)?;
-            if actual_speed != speed {
-                log::warn!(
-                    "Protocol speed {} kHz not supported, actual speed is {} kHz",
-                    speed,
-                    actual_speed
-                );
-            }
-        }
-
-        let mut permissions = Permissions::new();
-        if debugger_options.allow_erase_all {
-            permissions = permissions.allow_erase_all();
-        }
-
-        // Attach to the probe.
-        let target_session = if debugger_options.connect_under_reset {
-            target_probe.attach_under_reset(target_selector, permissions)?
-        } else {
-            target_probe
-                .attach(target_selector, permissions)
-                .map_err(|err| {
-                    anyhow!(
-                        "Error attaching to the probe: {:?}.\nTry the --connect-under-reset option",
-                        err
-                    )
-                })?
-        };
-
-        // Create an instance of the [`capstone::Capstone`] for disassembly capabilities.
-        let capstone = match target_session.architecture() {
-            probe_rs::Architecture::Arm => Capstone::new()
-                .arm()
-                .mode(armArchMode::Thumb)
-                .endian(Endian::Little)
-                .build()
-                .map_err(|err| anyhow!("Error creating Capstone disassembler: {:?}", err))?,
-            probe_rs::Architecture::Riscv => Capstone::new()
-                .riscv()
-                .mode(riscvArchMode::RiscV32)
-                .endian(Endian::Little)
-                .build()
-                .map_err(|err| anyhow!("Error creating Capstone disassembler: {:?}", err))?,
-        };
-
-        // Change the current working directory if `debugger_options.cwd` is `Some(T)`.
-        if let Some(new_cwd) = debugger_options.cwd.clone() {
-            set_current_dir(new_cwd.as_path()).map_err(|err| {
-                anyhow!(
-                    "Failed to set current working directory to: {:?}, {:?}",
-                    new_cwd,
-                    err
-                )
-            })?;
-        };
-
-        // TODO: We currently only allow a single core & binary to be specified in [DebuggerOptions]. When this is extended to support multicore, the following should initialize [DebugInfo] and [VariableCache] for each available core.
-        // Configure the [DebugInfo].
-        let debug_infos = vec![
-            if let Some(binary_path) = &debugger_options.program_binary {
-                DebugInfo::from_file(binary_path)
-                    .map_err(|error| DebuggerError::Other(anyhow!(error)))?
-            } else {
-                return Err(anyhow!(
-                    "Please provide a valid `program_binary` for this debug session"
-                )
-                .into());
-            },
-        ];
-
-        // Configure the [VariableCache].
-        let stack_frames = target_session.list_cores()
-            .iter()
-            .map(|(core_id, core_type)| {
-                log::debug!(
-                    "Preparing stack frame variable cache for DebugSession and CoreData for core #{} of type: {:?}",
-                    core_id,
-                    core_type
-                );
-                Vec::<probe_rs::debug::StackFrame>::new()
-            }).collect();
-
-        // Prepare the breakpoint cache
-        let breakpoints = target_session.list_cores()
-            .iter()
-            .map(|(core_id, core_type)| {
-                log::debug!(
-                    "Preparing breakpoint cache for DebugSession and CoreData for core #{} of type: {:?}",
-                    core_id,
-                    core_type
-                );
-                Vec::<ActiveBreakpoint>::new()
-            }).collect();
-
-        Ok(DebugSession {
-            session: target_session,
-            capstone,
-            debug_infos,
-            stack_frames,
-            breakpoints,
-            rtt_connection: None,
-        })
-    }
-
-    pub fn attach_core(&mut self, core_index: usize) -> Result<CoreData, DebuggerError> {
-        let target_name = self.session.target().name.clone();
-        // Do a 'light weight'(just get references to existing data structures) attach to the core and return relevant debug data.
-        match self.session.core(core_index) {
-            Ok(target_core) => Ok(CoreData {
-                target_core,
-                target_name: format!("{}-{}", core_index, target_name),
-                debug_info: self.debug_infos.get(core_index).ok_or_else(|| {
-                    DebuggerError::Other(anyhow!(
-                        "No available `DebugInfo` for core # {}",
-                        core_index
-                    ))
-                })?,
-                stack_frames: self.stack_frames.get_mut(core_index).ok_or_else(|| {
-                    DebuggerError::Other(anyhow!(
-                        "StackFrame cache was not correctly configured for core # {}",
-                        core_index
-                    ))
-                })?,
-                capstone: &self.capstone,
-                breakpoints: self.breakpoints.get_mut(core_index).ok_or_else(|| {
-                    DebuggerError::Other(anyhow!(
-                        "ActiveBreakpoint cache was not correctly configured for core # {}",
-                        core_index
-                    ))
-                })?,
-                rtt_connection: &mut self.rtt_connection,
-            }),
-            Err(_) => Err(DebuggerError::UnableToOpenProbe(Some(
-                "No core at the specified index.",
-            ))),
-        }
-    }
-}
-
-/// Manage the active RTT target for a specific DebugSession, as well as provide methods to reliably move RTT from target, through the debug_adapter, to the client.
-pub(crate) struct RttConnection {
-    /// The connection to RTT on the target
-    target_rtt: rtt::RttActiveTarget,
-    /// Some status fields and methods to ensure continuity in flow of data from target to debugger to client.
-    debugger_rtt_channels: Vec<DebuggerRttChannel>,
-}
-
-impl RttConnection {
-    /// Polls all the available channels for data and transmits data to the client.
-    /// If at least one channel had data, then return a `true` status.
-    pub fn process_rtt_data<P: ProtocolAdapter>(
-        &mut self,
-        debug_adapter: &mut DebugAdapter<P>,
-        target_core: &mut Core,
-    ) -> bool {
-        let mut at_least_one_channel_had_data = false;
-        for debugger_rtt_channel in self.debugger_rtt_channels.iter_mut() {
-            if debugger_rtt_channel.poll_rtt_data(target_core, debug_adapter, &mut self.target_rtt)
-            {
-                at_least_one_channel_had_data = true;
-            }
-        }
-        at_least_one_channel_had_data
-    }
-}
-
-pub(crate) struct DebuggerRttChannel {
-    pub(crate) channel_number: usize,
-    // We will not poll target RTT channels until we have confirmation from the client that the output window has been opened.
-    pub(crate) has_client_window: bool,
-}
-impl DebuggerRttChannel {
-    /// Poll and retrieve data from the target, and send it to the client, depending on the state of `hasClientWindow`.
-    /// Doing this selectively ensures that we don't pull data from target buffers until we have an output window, and also helps us drain buffers after the target has entered a `is_halted` state.
-    /// Errors will be reported back to the `debug_adapter`, and the return `bool` value indicates whether there was available data that was processed.
-    pub(crate) fn poll_rtt_data<P: ProtocolAdapter>(
-        &mut self,
-        core: &mut Core,
-        debug_adapter: &mut DebugAdapter<P>,
-        rtt_target: &mut rtt::RttActiveTarget,
-    ) -> bool {
-        if self.has_client_window {
-            rtt_target
-                .active_channels
-                .iter_mut()
-                .find(|active_channel| {
-                    if let Some(channel_number) = active_channel.number() {
-                        channel_number == self.channel_number
-                    } else {
-                        false
-                    }
-                })
-                .and_then(|rtt_channel| {
-                    rtt_channel.get_rtt_data(core, rtt_target.defmt_state.as_ref())
-                })
-                .and_then(|(channel_number, channel_data)| {
-                    if debug_adapter
-                        .rtt_output(channel_number.parse::<usize>().unwrap_or(0), channel_data)
-                    {
-                        Some(true)
-                    } else {
-                        None
-                    }
-                })
-                .is_some()
-        } else {
-            false
-        }
-    }
-}
-
-/// [CoreData] provides handles to various data structures required to debug a single instance of a core. The actual state is stored in [SessionData].
-///
-/// Usage: To get access to this structure please use the [DebugSession::attach_core] method. Please keep access/locks to this to a minumum duration.
-pub struct CoreData<'p> {
-    pub(crate) target_core: Core<'p>,
-    pub(crate) target_name: String,
-    pub(crate) debug_info: &'p DebugInfo,
-    pub(crate) stack_frames: &'p mut Vec<probe_rs::debug::StackFrame>,
-    pub(crate) capstone: &'p Capstone,
-    pub(crate) breakpoints: &'p mut Vec<ActiveBreakpoint>,
-    pub(crate) rtt_connection: &'p mut Option<RttConnection>,
-}
-
-impl<'p> CoreData<'p> {
-    /// Search available [StackFrame]'s for the given `id`
-    pub(crate) fn get_stackframe(&'p self, id: i64) -> Option<&'p probe_rs::debug::StackFrame> {
-        self.stack_frames
-            .iter()
-            .find(|stack_frame| stack_frame.id == id)
-    }
-
-    /// Confirm RTT initialization on the target, and use the RTT channel configurations to initialize the output windows on the DAP Client.
-    pub fn attach_to_rtt<P: ProtocolAdapter>(
-        &mut self,
-        debug_adapter: &mut DebugAdapter<P>,
-        target_memory_map: &[probe_rs::config::MemoryRegion],
-        program_binary: &std::path::Path,
-        rtt_config: &rtt::RttConfig,
-    ) -> Result<()> {
-        let mut debugger_rtt_channels: Vec<DebuggerRttChannel> = vec![];
-        match rtt::attach_to_rtt(
-            &mut self.target_core,
-            target_memory_map,
-            program_binary,
-            rtt_config,
-        ) {
-            Ok(target_rtt) => {
-                for any_channel in target_rtt.active_channels.iter() {
-                    if let Some(up_channel) = &any_channel.up_channel {
-                        debugger_rtt_channels.push(DebuggerRttChannel {
-                            channel_number: up_channel.number(),
-                            // This value will eventually be set to true by a VSCode client request "rttWindowOpened"
-                            has_client_window: false,
-                        });
-                        debug_adapter.rtt_window(
-                            up_channel.number(),
-                            any_channel.channel_name.clone(),
-                            any_channel.data_format,
-                        );
-                    }
-                }
-                *self.rtt_connection = Some(RttConnection {
-                    target_rtt,
-                    debugger_rtt_channels,
-                });
-            }
-            Err(_error) => {
-                log::warn!("Failed to initalize RTT. Will try again on the next request... ");
-            }
-        };
-        Ok(())
-    }
-
-    /// Set a single breakpoint in target configuration as well as [`CoreData::breakpoints`]
-    pub(crate) fn set_breakpoint(
-        &mut self,
-        address: u32,
-        breakpoint_type: BreakpointType,
-    ) -> Result<(), DebuggerError> {
-        self.target_core
-            .set_hw_breakpoint(address)
-            .map_err(DebuggerError::ProbeRs)?;
-        self.breakpoints.push(ActiveBreakpoint {
-            breakpoint_type,
-            breakpoint_address: address,
-        });
-        Ok(())
-    }
-
-    /// Clear a single breakpoint from target configuration as well as [`CoreData::breakpoints`]
-    pub(crate) fn clear_breakpoint(&mut self, address: u32) -> Result<()> {
-        self.target_core
-            .clear_hw_breakpoint(address)
-            .map_err(DebuggerError::ProbeRs)?;
-        let mut breakpoint_position: Option<usize> = None;
-        for (position, active_breakpoint) in self.breakpoints.iter().enumerate() {
-            if active_breakpoint.breakpoint_address == address {
-                breakpoint_position = Some(position);
-                break;
-            }
-        }
-        if let Some(breakpoint_position) = breakpoint_position {
-            self.breakpoints.remove(breakpoint_position as usize);
-        }
-        Ok(())
-    }
-
-    /// Clear all breakpoints of a specified [`BreakpointType`]. Affects target configuration as well as [`CoreData::breakpoints`]
-    pub(crate) fn clear_breakpoints(&mut self, breakpoint_type: BreakpointType) -> Result<()> {
-        let target_breakpoints = self
-            .breakpoints
-            .iter()
-            .filter(|breakpoint| breakpoint.breakpoint_type == breakpoint_type)
-            .map(|breakpoint| breakpoint.breakpoint_address)
-            .collect::<Vec<u32>>();
-        for breakpoint in target_breakpoints {
-            self.clear_breakpoint(breakpoint).ok();
-        }
-        Ok(())
-    }
-}
-
 #[derive(Debug)]
 /// The `DebuggerStatus` is used to control how the Debugger::debug_session() decides if it should respond to DAP Client requests such as `Terminate`, `Disconnect`, and `Reset`, as well as how to repond to unrecoverable errors during a debug session interacting with a target session.
 pub(crate) enum DebuggerStatus {
@@ -652,19 +57,22 @@ pub(crate) enum DebuggerStatus {
 /// - In this case, the management (start and stop) of the server process is the responsibility of the user. e.g.
 ///   - `probe-rs-debug --debug --port <IP port number> <other options>` : Uses TCP Sockets to the defined IP port number to service DAP requests.
 pub struct Debugger {
-    debugger_options: DebuggerOptions,
+    config: configuration::SessionConfig,
 }
 
 impl Debugger {
-    pub fn new(debugger_options: DebuggerOptions) -> Self {
-        // Define all the commands supported by the debugger.
-        // TODO: There is a lot of repetitive code here, and a great opportunity for macros.
-        Self { debugger_options }
+    pub fn new(port: Option<u16>) -> Self {
+        Self {
+            config: configuration::SessionConfig {
+                port,
+                ..Default::default()
+            },
+        }
     }
 
     pub(crate) fn process_next_request<P: ProtocolAdapter>(
         &mut self,
-        session_data: &mut DebugSession,
+        session_data: &mut session_data::SessionData,
         debug_adapter: &mut DebugAdapter<P>,
     ) -> Result<DebuggerStatus, DebuggerError> {
         let request = debug_adapter.listen_for_request()?;
@@ -683,7 +91,7 @@ impl Debugger {
                     CoreStatus::Unknown => Ok(DebuggerStatus::ContinueSession), // Don't do anything until we know VSCode's startup sequence is complete, and changes this to either Halted or Running.
                     CoreStatus::Halted(_) => {
                         // Make sure the RTT buffers are drained.
-                        match session_data.attach_core(self.debugger_options.core_index) {
+                        match session_data.attach_core(self.config.core_index) {
                             Ok(mut core_data) => {
                                 if let Some(rtt_active_target) = &mut core_data.rtt_connection {
                                     rtt_active_target.process_rtt_data(
@@ -704,23 +112,22 @@ impl Debugger {
                     }
                     _other => {
                         let mut received_rtt_data = false;
-                        let mut core_data =
-                            match session_data.attach_core(self.debugger_options.core_index) {
-                                Ok(mut core_data) => {
-                                    // Use every opportunity to poll the RTT channels for data
-                                    if let Some(rtt_active_target) = &mut core_data.rtt_connection {
-                                        received_rtt_data = rtt_active_target.process_rtt_data(
-                                            debug_adapter,
-                                            &mut core_data.target_core,
-                                        );
-                                    }
-                                    core_data
+                        let mut core_data = match session_data.attach_core(self.config.core_index) {
+                            Ok(mut core_data) => {
+                                // Use every opportunity to poll the RTT channels for data
+                                if let Some(rtt_active_target) = &mut core_data.rtt_connection {
+                                    received_rtt_data = rtt_active_target.process_rtt_data(
+                                        debug_adapter,
+                                        &mut core_data.target_core,
+                                    );
                                 }
-                                Err(error) => {
-                                    let _ = debug_adapter.send_error_response(&error)?;
-                                    return Err(error);
-                                }
-                            };
+                                core_data
+                            }
+                            Err(error) => {
+                                let _ = debug_adapter.send_error_response(&error)?;
+                                return Err(error);
+                            }
+                        };
 
                         // Check and update the core status.
                         let new_status = match core_data.target_core.status() {
@@ -790,8 +197,7 @@ impl Debugger {
             }
             Some(request) => {
                 // First, attach to the core.
-                let mut core_data = match session_data.attach_core(self.debugger_options.core_index)
-                {
+                let mut core_data = match session_data.attach_core(self.config.core_index) {
                     Ok(core_data) => core_data,
                     Err(error) => {
                         let failed_command = request.command.clone();
@@ -1114,17 +520,17 @@ impl Debugger {
         match get_arguments(&launch_attach_request) {
             Ok(arguments) => {
                 if requested_target_session_type.is_some() {
-                    self.debugger_options = DebuggerOptions { ..arguments };
+                    self.config = configuration::SessionConfig { ..arguments };
                     if matches!(
                         requested_target_session_type,
                         Some(TargetSessionType::AttachRequest)
                     ) {
                         // Since VSCode doesn't do field validation checks for relationships in launch.json request types, check it here.
-                        if self.debugger_options.flashing_enabled
-                            || self.debugger_options.reset_after_flashing
-                            || self.debugger_options.halt_after_reset
-                            || self.debugger_options.full_chip_erase
-                            || self.debugger_options.restore_unwritten_bytes
+                        if self.config.flashing_enabled
+                            || self.config.reset_after_flashing
+                            || self.config.halt_after_reset
+                            || self.config.full_chip_erase
+                            || self.config.restore_unwritten_bytes
                         {
                             debug_adapter.send_response::<()>(
                                         launch_attach_request,
@@ -1138,17 +544,14 @@ impl Debugger {
                     }
                 }
                 debug_adapter.set_console_log_level(
-                    self.debugger_options
-                        .console_log_level
-                        .unwrap_or(ConsoleLog::Error),
+                    self.config.console_log_level.unwrap_or(ConsoleLog::Error),
                 );
                 // Update the `cwd` and `program_binary`.
-                self.debugger_options
-                    .validate_and_update_cwd(self.debugger_options.cwd.clone());
+                self.config.validate_and_update_cwd(self.config.cwd.clone());
                 // Update the `program_binary` and validate that the file exists.
-                self.debugger_options.program_binary = match self
-                    .debugger_options
-                    .qualify_and_update_os_file_path(self.debugger_options.program_binary.clone())
+                self.config.program_binary = match self
+                    .config
+                    .qualify_and_update_os_file_path(self.config.program_binary.clone())
                 {
                     Ok(program_binary) => {
                         if !program_binary.is_file() {
@@ -1180,9 +583,9 @@ impl Debugger {
                 };
                 // Update the `svd_file` and validate that the file exists.
                 // If there is a problem with this file, warn the user and continue with the session.
-                self.debugger_options.svd_file = match self
-                    .debugger_options
-                    .qualify_and_update_os_file_path(self.debugger_options.svd_file.clone())
+                self.config.svd_file = match self
+                    .config
+                    .qualify_and_update_os_file_path(self.config.svd_file.clone())
                 {
                     Ok(svd_file) => {
                         if !svd_file.is_file() {
@@ -1204,7 +607,10 @@ impl Debugger {
                 debug_adapter.send_response::<()>(launch_attach_request, Ok(None))?;
             }
             Err(error) => {
-                let error_message = format!("Could not derive DebuggerOptions from request '{}', with arguments {:?}\n{:?} ", launch_attach_request.command, launch_attach_request.arguments, error);
+                let error_message = format!(
+                    "Could not derive SessionConfig from request '{}', with arguments {:?}\n{:?} ",
+                    launch_attach_request.command, launch_attach_request.arguments, error
+                );
                 debug_adapter.send_response::<()>(
                     launch_attach_request,
                     Err(DebuggerError::Other(anyhow!(error_message.clone()))),
@@ -1214,19 +620,19 @@ impl Debugger {
             }
         };
 
-        let mut session_data = match DebugSession::new(&self.debugger_options) {
+        let mut session_data = match session_data::SessionData::new(&self.config) {
             Ok(session_data) => session_data,
             Err(error) => {
                 debug_adapter.send_error_response(&error)?;
                 return Err(error);
             }
         };
-        debug_adapter.halt_after_reset = self.debugger_options.halt_after_reset;
+        debug_adapter.halt_after_reset = self.config.halt_after_reset;
 
         // Do the flashing.
         {
-            if self.debugger_options.flashing_enabled {
-                let path_to_elf = match self.debugger_options.program_binary.clone() {
+            if self.config.flashing_enabled {
+                let path_to_elf = match self.config.program_binary.clone() {
                     Some(program_binary) => program_binary,
                     None => {
                         let err = DebuggerError::Other(anyhow!(
@@ -1244,9 +650,8 @@ impl Debugger {
                 let progress_id = debug_adapter.start_progress("Flashing device", None).ok();
 
                 let mut download_options = DownloadOptions::default();
-                download_options.keep_unwritten_bytes =
-                    self.debugger_options.restore_unwritten_bytes;
-                download_options.do_chip_erase = self.debugger_options.full_chip_erase;
+                download_options.keep_unwritten_bytes = self.config.restore_unwritten_bytes;
+                download_options.do_chip_erase = self.config.full_chip_erase;
                 let flash_result = {
                     let rc_debug_adapter = Rc::new(RefCell::new(debug_adapter));
                     let rc_debug_adapter_clone = rc_debug_adapter.clone();
@@ -1457,10 +862,10 @@ impl Debugger {
         // This is the first attach to the requested core. If this one works, all subsequent ones will be no-op requests for a Core reference. Do NOT hold onto this reference for the duration of the session ... that is why this code is in a block of its own.
         {
             // First, attach to the core
-            let mut core_data = match session_data.attach_core(self.debugger_options.core_index) {
+            let mut core_data = match session_data.attach_core(self.config.core_index) {
                 Ok(mut core_data) => {
                     // Immediately after attaching, halt the core, so that we can finish initalization without bumping into user code.
-                    // Depending on supplied `debugger_options`, the core will be restarted at the end of initialization in the `configuration_done` request.
+                    // Depending on supplied `config`, the core will be restarted at the end of initialization in the `configuration_done` request.
                     match halt_core(&mut core_data.target_core) {
                         Ok(_) => {}
                         Err(error) => {
@@ -1476,8 +881,7 @@ impl Debugger {
                 }
             };
 
-            if self.debugger_options.flashing_enabled && self.debugger_options.reset_after_flashing
-            {
+            if self.config.flashing_enabled && self.config.reset_after_flashing {
                 debug_adapter
                     .restart(&mut core_data, None)
                     .context("Failed to restart core")?;
@@ -1503,21 +907,20 @@ impl Debugger {
             match self.process_next_request(&mut session_data, &mut debug_adapter) {
                 Ok(DebuggerStatus::ContinueSession) => {
                     // Validate and if necessary, initialize the RTT structure.
-                    if self.debugger_options.rtt.enabled
+                    if self.config.rtt.enabled
                         && session_data.rtt_connection.is_none()
                         && !(debug_adapter.last_known_status == CoreStatus::Unknown
                             || debug_adapter.last_known_status.is_halted())
                     // Do not attempt this until we have processed the MSDAP request for "configurationDone" ...
                     {
                         let target_memory_map = session_data.session.target().memory_map.clone();
-                        let mut core_data =
-                            match session_data.attach_core(self.debugger_options.core_index) {
-                                Ok(core_data) => core_data,
-                                Err(error) => {
-                                    debug_adapter.send_error_response(&error)?;
-                                    return Err(error);
-                                }
-                            };
+                        let mut core_data = match session_data.attach_core(self.config.core_index) {
+                            Ok(core_data) => core_data,
+                            Err(error) => {
+                                debug_adapter.send_error_response(&error)?;
+                                return Err(error);
+                            }
+                        };
                         log::info!("Attempting to initialize the RTT.");
                         // RTT can only be initialized if the target application has been allowed to run to the point where it does the RTT initialization.
                         // If the target halts before it processes this code, then this RTT intialization silently fails, and will try again later ...
@@ -1527,8 +930,8 @@ impl Debugger {
                         core_data.attach_to_rtt(
                             &mut debug_adapter,
                             &target_memory_map,
-                            self.debugger_options.program_binary.as_ref().unwrap(),
-                            &self.debugger_options.rtt,
+                            self.config.program_binary.as_ref().unwrap(),
+                            &self.config.rtt,
                         )?;
                     }
                 }
@@ -1587,16 +990,16 @@ pub fn list_supported_chips() -> Result<()> {
     Ok(())
 }
 
-pub fn debug(debugger_options: DebuggerOptions, vscode: bool) -> Result<()> {
+pub fn debug(port: Option<u16>, vscode: bool) -> Result<()> {
     let program_name = clap::crate_name!();
 
-    let mut debugger = Debugger::new(debugger_options);
+    let mut debugger = Debugger::new(port);
 
     println!(
         "{} CONSOLE: Starting as a DAP Protocol server",
         &program_name
     );
-    match &debugger.debugger_options.port.clone() {
+    match &debugger.config.port.clone() {
         Some(port) => {
             let addr = std::net::SocketAddr::new(
                 std::net::IpAddr::V4(Ipv4Addr::LOCALHOST),
